solve fitur error

+++ sita/sita/src/main/java/protensi/sita/service/MahasiswaService.java
@@ -1,10 +1,7 @@
 package protensi.sita.service;
 
-import protensi.sita.model.EnumRole;
 import protensi.sita.model.MahasiswaModel;
-import java.util.List;
-
 
 public interface MahasiswaService {
-    MahasiswaModel findMahasiswaById(Integer idUser);
+    MahasiswaModel findMahasiswaById(Long idUser);
 }
